remove entity crate
