+++ dev/server/grpc/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-relentless-dev-server-grpc-entity = { path = "./entity" }
 prost = "0.13.4"
 prost-types = "0.13.4"
 tonic = "0.12.3"
@@ -21,3 +20,6 @@ tracing-subscriber = "0.3.19"
 
 num = "0.4.3"
 thiserror = "2.0.9"
+
+[build-dependencies]
+tonic-build = "0.12.3"
