+++ dev/server/grpc/examples/echo.rs
@@ -1,5 +1,5 @@
 use prost_types::{Any, Value};
-use relentless_dev_server_grpc_entity::echo_pb::echo_client::EchoClient;
+use relentless_dev_server_grpc::service::echo::pb::echo_client::EchoClient;
 use tonic::Request;
 
 #[tokio::main]
