+++ dev/server/grpc/src/service.rs
@@ -1,8 +1,7 @@
+use counter::pb::counter_server::CounterServer;
+use echo::pb::echo_server::EchoServer;
+use helloworld::pb::greeter_server::GreeterServer;
 use num::BigInt;
-use relentless_dev_server_grpc_entity::{
-    counter_pb::counter_server::CounterServer, echo_pb::echo_server::EchoServer,
-    helloworld_pb::greeter_server::GreeterServer,
-};
 use tonic::transport::{server::Router, Server};
 use tonic_health::{pb::health_server::HealthServer, server::HealthService};
 use tower::layer::util::Identity;
@@ -13,6 +12,8 @@ pub mod counter;
 pub mod echo;
 pub mod helloworld;
 
+pub const FILE_DESCRIPTOR_SET: &[u8] = tonic::include_file_descriptor_set!("file_descriptor");
+
 pub async fn app(env: Env) -> Router<Identity> {
     app_with(env, 0.into()).await
 }
@@ -21,11 +22,12 @@ pub async fn app_with(env: Env, initial_count: BigInt) -> Router<Identity> {
     health_reporter.set_serving::<HealthServer<HealthService>>().await;
     health_reporter.set_serving::<GreeterServer<helloworld::MyGreeter>>().await;
     health_reporter.set_serving::<CounterServer<counter::CounterImpl>>().await;
+    health_reporter.set_serving::<EchoServer<echo::EchoImpl>>().await;
 
     let reflection_service = tonic_reflection::server::Builder::configure()
         .register_encoded_file_descriptor_set(tonic_health::pb::FILE_DESCRIPTOR_SET)
         .register_encoded_file_descriptor_set(tonic_reflection::pb::v1::FILE_DESCRIPTOR_SET)
-        .register_encoded_file_descriptor_set(relentless_dev_server_grpc_entity::FILE_DESCRIPTOR_SET)
+        .register_encoded_file_descriptor_set(FILE_DESCRIPTOR_SET)
         .build_v1()
         .unwrap();
