+++ dev/server/grpc/src/service/echo.rs
@@ -1,12 +1,15 @@
 use prost_types::{Any, Value};
-use relentless_dev_server_grpc_entity::echo_pb::echo_server::Echo;
 use tonic::{Request, Response, Status};
+
+pub mod pb {
+    tonic::include_proto!("echo");
+}
 
 #[derive(Debug, Default)]
 pub struct EchoImpl;
 
 #[tonic::async_trait]
-impl Echo for EchoImpl {
+impl pb::echo_server::Echo for EchoImpl {
     #[tracing::instrument(ret)]
     async fn echo(&self, request: Request<Any>) -> Result<Response<Any>, Status> {
         let value = request.into_inner();
